Cuts-from-proofs: refactor the interface
Move sparse matrix representation to a separate file, and remove
dependency on ArithLogic
